Synced credit score update with profile

+++ src/main/java/com/example/demoApp/demo/dao/Profile.java
@@ -13,7 +13,7 @@ public class Profile {
     @Id
     private String mobileNumber;
     private String orgName;
-    private int creditScore;
+    private double creditScore;
     private int amountToGive;
     private int amountToPay;
 }
